feat: baseline compilation node flags
